fix(test): resolve type mismatches in infra tests
- Fix float32/float64 type mismatch in hash consistent hashing test
- Fix int vs int64 type mismatch in mathx unstable test
- Fix int32 vs int type mismatch in mr mapreduce test
- Fix assert.ErrorIs argument order in search tree test

+++ infra/mathx/unstable_test.go
@@ -31,7 +31,7 @@ func TestUnstable_AroundIntLarge(t *testing.T) {
 	unstable := NewUnstable(5)
 	for range 1000 {
 		val := unstable.AroundInt(target)
-		assert.LessOrEqual(t, 0, val)
+		assert.LessOrEqual(t, int64(0), val)
 		assert.LessOrEqual(t, val, 2*target)
 	}
 }
